Move maze generation to sketchlib
